+++ event/parsing/totalBootTimeParser.go
@@ -5,14 +5,15 @@
 package parsing
 
 import (
+	"errors"
 	"regexp"
 	"time"
 
 	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
+	"github.com/goph/emperror"
 	"github.com/xmidt-org/glaukos/event/queue"
 	"github.com/xmidt-org/themis/xlog"
-	"github.com/xmidt-org/wrp-go/v3"
 )
 
 // TotalBootTimeParser takes fully-manageabke events and calculates the reboot time of a device by getting the last
@@ -59,29 +60,30 @@ func (b *TotalBootTimeParser) calculateRestartTime(wrpWithTime queue.WrpWithTime
 		return -1, err
 	}
 
-	previousBootTime := int64(0)
-	latestRebootEvent := int64(0)
-
 	// Get events from codex pertaining to this device id.
 	events := b.Client.GetEvents(deviceID)
+	latestPreviousEvent := Event{}
 
-	// Find the previous boot-time and make sure that the boot time we have is the latest one.
+	// Go through events to find reboot-pending event with the boot-time of the previous session
 	for _, event := range events {
-		if previousBootTime, newBootTimeFound, err := getPreviousBootTime(event, previousBootTime, bootTimeInt, msg); err != nil {
+		if latestPreviousEvent, err = checkLatestPreviousEvent(event, latestPreviousEvent, bootTimeInt, rebootRegex); err != nil {
 			level.Error(b.Logger).Log(xlog.ErrorKey(), err)
-			if previousBootTime < 0 {
+			if errors.Is(err, errNewerBootTime) {
 				// Something is wrong with this event's boot time, we shouldn't continue.
 				b.Measures.UnparsableEventsCount.With(ParserLabel, b.Label, ReasonLabel, eventBootTimeErr).Add(1.0)
 				return -1, err
 			}
-		} else if newBootTimeFound {
-			latestRebootEvent, _ = getLastRebootEvent(event, latestRebootEvent)
 		}
+	}
+
+	if valid, err := isEventValid(latestPreviousEvent, rebootRegex, time.Now); !valid {
+		level.Error(b.Logger).Log(xlog.ErrorKey(), err)
+		return -1, err
 	}
 
 	// boot time calculation
 	// Event birthdate is saved in unix nanoseconds, so we must first convert it to a unix time using nanoseconds.
-	restartTime := wrpWithTime.Beginning.Sub(time.Unix(0, latestRebootEvent)).Seconds()
+	restartTime := wrpWithTime.Beginning.Sub(time.Unix(0, latestPreviousEvent.BirthDate)).Seconds()
 
 	if restartTime <= 0 {
 		err = errInvalidRestartTime
@@ -89,39 +91,80 @@ func (b *TotalBootTimeParser) calculateRestartTime(wrpWithTime queue.WrpWithTime
 		return -1, err
 	}
 
-	if latestRebootEvent != 0 {
-		return restartTime, nil
-	}
-
-	err = errRestartTime
-	level.Error(b.Logger).Log(xlog.ErrorKey(), err)
-	return -1, err
-
+	return restartTime, nil
 }
 
-// Get the latest previous boot time. Returns the latest previous boot time found, a boolean indicating if a more recent boot-time
-// has been found, and a possible error
-func getPreviousBootTime(e Event, previousBootTime int64, latestBootTime int64, currentMsg wrp.Message) (int64, bool, error) {
+// sees if this event is part of the most recent previous session
+func checkLatestPreviousEvent(e Event, previousEventTracked Event, latestBootTime int64, eventType *regexp.Regexp) (Event, error) {
 	eventBootTimeInt, err := GetEventBootTime(e)
+	previousEventBootTime, _ := GetEventBootTime(previousEventTracked)
+
 	if err != nil {
-		return previousBootTime, false, err
+		return previousEventTracked, err
 	}
 
 	if eventBootTimeInt > latestBootTime {
-		return -1, false, errNewerBootTime
+		return Event{}, errNewerBootTime
 	}
 
-	if eventBootTimeInt > previousBootTime && e.TransactionUUID != currentMsg.TransactionUUID {
-		return eventBootTimeInt, true, nil
+	// If this event has a boot time greater than what we've seen so far
+	// but still less than the current boot-time, it means that this event
+	// is part of a more recent previous cycle.
+	if eventBootTimeInt > previousEventBootTime && eventBootTimeInt < latestBootTime {
+		return e, nil
 	}
 
-	return previousBootTime, false, nil
+	// If the event is of the type we are looking for and has the same boot-time as the
+	// newest previous boot time, return this event.
+	if eventBootTimeInt == previousEventBootTime && eventType.MatchString(e.Dest) {
+		// If the previously tracked event doesn't match the event we're looking for,
+		// return the current event. If both events match the type of event we are looking for,
+		// compare the birthdates and return the one with the more recent birthdate.
+		if !eventType.MatchString(previousEventTracked.Dest) || e.BirthDate > previousEventTracked.BirthDate {
+			return e, nil
+		}
+	}
+	return previousEventTracked, nil
 }
 
-// get the last reboot-pending event
-func getLastRebootEvent(e Event, latestRebootBirthDate int64) (int64, error) {
-	if rebootRegex.MatchString(e.Dest) && e.BirthDate > latestRebootBirthDate {
-		return e.BirthDate, nil
+// Sees if a date is within a certain time frame.
+// PastBuffer should be a positive duration.
+func isDateValid(currTime func() time.Time, pastBuffer time.Duration, futureBuffer time.Duration, date time.Time) bool {
+	if date.Before(time.Unix(0, 0)) || date.Equal(time.Unix(0, 0)) {
+		return false
 	}
-	return latestRebootBirthDate, nil
+
+	if pastBuffer.Seconds() < 0 {
+		pastBuffer = -1 * pastBuffer
+	}
+
+	now := currTime()
+	pastTime := now.Add(-1 * pastBuffer)
+	futureTime := now.Add(futureBuffer)
+
+	return (pastTime.Before(date) || pastTime.Equal(date)) && (futureTime.Equal(date) || futureTime.After(date))
+}
+
+func isEventValid(event Event, expectedType *regexp.Regexp, currTime func() time.Time) (bool, error) {
+	// see if event found matches expected event type
+	if !expectedType.MatchString(event.Dest) {
+		return false, emperror.WrapWith(errors.New("unexpected event type"), errRestartTime.Error(), "expected type", expectedType.String())
+	}
+
+	// check if boot-time is valid
+	bootTime, err := GetEventBootTime(event)
+	if bootTime <= 0 {
+		return false, emperror.WrapWith(errors.New("Unable to get boot-time"), errRestartTime.Error(), "parsing error", err)
+	}
+
+	if !isDateValid(currTime, 12*time.Hour, time.Hour, time.Unix(bootTime, 0)) {
+		return false, emperror.WrapWith(errors.New("invalid boot-time"), errRestartTime.Error())
+	}
+
+	// see if birthdate is valid
+	if !isDateValid(currTime, 12*time.Hour, time.Hour, time.Unix(0, event.BirthDate)) {
+		return false, emperror.WrapWith(errors.New("invalid birthdate"), errRestartTime.Error())
+	}
+
+	return true, nil
 }